requirements.txt
asgiref==3.4.1
black==21.11b1
certifi==2021.10.8
charset-normalizer==2.0.8
click==8.0.3
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.9
django-admin-interface==0.18.2
django-colorfield==0.4.5
django-filter==21.1
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-js-asset==1.2.2
django-model-utils==4.2.0
django-mptt==0.13.4
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
drf-yasg==1.20.0
idna==3.3
importlib-metadata==4.8.2
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
Markdown==3.3.6
MarkupSafe==2.0.1
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.9.0
platformdirs==2.4.0
psycopg2-binary==2.9.2
PyJWT==2.3.0
pyparsing==3.0.6
pytz==2021.3
regex==2021.11.10
requests==2.26.0
ruamel.yaml==0.17.17
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
tomli==1.2.2
typing_extensions==4.0.0
uritemplate==4.1.1
urllib3==1.26.7
zipp==3.6.0
